requirements.txt
numpy>=1.26
scipy>=1.12
matplotlib>=3.8
pydantic>=2.5
typer>=0.12
rich>=13.0
pytest>=8
h5py>=3.11
cupy-cuda12x>=13.0; platform_system=="Windows"
plotly>=5.20